Implement Enhancement (Issue #129)
Add getRandomAI() method to VDomServer.java, which checks if the "Random AI" value is in the players list, then replaces that with a different AI not currently present in the game.

Add "Random AI" to player list in StartGameDialog as well.

Also, I work with IntellJ, so I added an ignore for IntelliJ-specific files.

git-svn-id: http://androminion.googlecode.com/svn/trunk@323 479f531a-0727-ffee-c1d2-1d776a0eae4d