Refactoring of the way the piles are selected and added to the game.  This only includes random sets and pre-defined sets; all sets from Dominion Shuffle are entered the same way as before.
The CardSet class contains the random and pre-defined sets from the game, and methods for interacting with them.  Using the GameType, a set is returned to the Game class, and the piles are added from there.  All card selections are now done in the CardSet class as well, so the Game class just needs to get the CardSet object, add the cards to piles, and if applicable, add the Bane card as well.

git-svn-id: http://androminion.googlecode.com/svn/trunk@366 479f531a-0727-ffee-c1d2-1d776a0eae4d