Added view for non-supply cards from Dark Ages (similar to prize view). Implemented Bandit Camp and Spoils (non-supply).
Notes:

Spoils is immediately returned to its non-supply pile upon being played, yet it still shows up in the played pile (for the moment). It might be beneficial to make the card semi-transparent (like a duration card?) to indicate that it's not really there any more (or it could just not show up in the played pile at all). Horn of Plenty will not count Spoils when determining the number of unique cards in play, as Spoils has already been returned to the pile.

Mercenary and Madman should eventually exist in the non-supply pile as well. Ruins are technically in the supply when in play, and should be placed in the Supply view. Shelters don't need piles, as far as I know.

git-svn-id: http://androminion.googlecode.com/svn/branches/DarkAges@650 479f531a-0727-ffee-c1d2-1d776a0eae4d