Some very preliminary work on multiplayer
The (first) goal to get multiplayer working is to break the dependency of
IndirectPlayer and RemotePlayer on the android framework, so we can move them
into the vdom code.  To do that, I'm making the client/server API use the
Options enums instead of literal strings, and putting GameTable and/or
GameActivity in charge of producing literal strings for each card.

git-svn-id: http://androminion.googlecode.com/svn/branches/Multiplayer@806 479f531a-0727-ffee-c1d2-1d776a0eae4d