Moved Strings.java to a place that fits better
The whole point is that we don't want to deal with strings in the server, so
it makes sense move the Strings class out of the server package.  In a
model-view-controller framework, it's part of the view, so it's going in the
UI package.  This also makes it more obvious when we're done with this part of
the multiplayer work: when we can remove all of the com.mehtank.androminion.*
imports.  Then we can safely move the files over into the vdom codebase, and
start actually implementing _new_ features that allow multiple remote players
to connect to the server.

git-svn-id: http://androminion.googlecode.com/svn/branches/Multiplayer@829 479f531a-0727-ffee-c1d2-1d776a0eae4d