Fixed Issue 605 (seemingly random crashes)
This happened because the AI player classes did not have the updated API for
Doctor and Journeyman (the one introduced to fix Issue 593).  So when an AI
tried to play one of those cards, the app crashed in a way that wasn't very
transparent to the user, and it seemed random.  It was immediately obvious what
the problem was upon looking at a log file, however.

My build script was only recompiling the classes that changed, and so it didn't
notice that the AI classes no longer matched the API.  I changed the build
script to clean the vdom lib before making a jar, so that this should not
happen again, though builds will take a little bit longer (not that much, as
compiling vdom_lib is pretty quick).

git-svn-id: http://androminion.googlecode.com/svn/branches/Multiplayer@851 479f531a-0727-ffee-c1d2-1d776a0eae4d