Removed dependency on Strings from RemotePlayer
If not for RemotePlayer's dependency on VDomServer, it could now be moved over
to the vdom code.  Instead,  we need to wait until VDomServer is ready to go
too (which shouldn't take long).  I still haven't tested this apk, however,
and I need to look more at the VDomServer code to better understand the
dependency between RemotePlayer and VDomServer.

git-svn-id: http://androminion.googlecode.com/svn/branches/Multiplayer@874 479f531a-0727-ffee-c1d2-1d776a0eae4d